% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFuncs.R
\name{fitLimma}
\alias{fitLimma}
\title{fit limma model}
\usage{
fitLimma(
  data,
  grp,
  contr = NULL,
  extra = NULL,
  pathway = NULL,
  method = "ls",
  plotHists = F,
  trend = F,
  subset = NULL
)
}
\arguments{
\item{data}{input data (feature * samples)}

\item{grp}{data group information}

\item{contr}{manually set contrast, default=NULL}
}
\description{
fit limma model
}
