% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFuncs.R
\name{QV}
\alias{QV}
\title{Calculate q-value. If reports Q-value error, adjusted p-value with BH adjustment will be returned}
\usage{
QV(pval)
}
\description{
Calculate q-value. If reports Q-value error, adjusted p-value with BH adjustment will be returned
}
