% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFuncs.R
\name{runGnetMulti}
\alias{runGnetMulti}
\title{multi elastic net regression}
\usage{
runGnetMulti(
  x,
  y,
  family = NULL,
  nrep = 100,
  alpha = 0.9,
  bootstrap = F,
  seed = 123,
  ...
)
}
\arguments{
\item{x}{data (feature * samples)}

\item{y}{label}

\item{family}{"binomial" or "gaussian" for y}

\item{nrep}{total number of elastic net regressions, default nrep=100}

\item{alpha}{elastic net mixing parameter, default alpha=0.9}

\item{bootstrap}{default bootstrap=T}

\item{seed}{random seed, default seed=123}
}
\description{
multi elastic net regression
}
