% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFuncs.R
\name{tscale}
\alias{tscale}
\title{scale data to mean=0 and sd=1}
\usage{
tscale(x, margin = 1)
}
\arguments{
\item{x}{data matrix}

\item{margin}{margin=1 for rows, margin=2 for columns. default margin=1}
}
\value{
A list for feature selection results.
}
\description{
scale data to mean=0 and sd=1
}
